+++ src/py/flwr/server/superlink/fleet/vce/vce_api.py
@@ -28,7 +28,7 @@ from typing import Callable, Dict, Optional
 
 from flwr.client.client_app import ClientApp, ClientAppException, LoadClientAppError
 from flwr.client.node_state import NodeState
-from flwr.client.process.utils import get_load_client_app_fn
+from flwr.client.clientapp.utils import get_load_client_app_fn
 from flwr.common.constant import (
     NUM_PARTITIONS_KEY,
     PARTITION_ID_KEY,
